Split apart video.py
