+++ frigate/video/__init__.py
@@ -1,0 +1,2 @@
+from .detect import *  # noqa: F403
+from .ffmpeg import *  # noqa: F403
